Updated Ninject code to support v6
