+++ packaging/nuget/NServiceBus.Ninject.nuspec
@@ -19,6 +19,7 @@
       <dependency id="Ninject" version="[3.2.2, 4.0.0)" />
       <dependency id="Ninject.Extensions.ContextPreservation" version="[3.2.0, 4.0.0)" />
       <dependency id="Ninject.Extensions.NamedScope" version="[3.2.0, 4.0.0)" />
+      <dependency id="Ninject.Extensions.ChildKernel" version="[3.2.0, 4.0.0)" />
     </dependencies>
   </metadata>
   <files>
