+++ src/NServiceBus.Ninject/NinjectBuilder.cs
@@ -13,7 +13,7 @@
         /// <summary>
         ///     Implementers need to new up a new container.
         /// </summary>
-        /// <param name="settings">The settings to check if an existing container exists.</param>
+        /// <param name="settings">The <paramref name="settings"/> to check if an existing container exists.</param>
         /// <returns>The new container wrapper.</returns>
         public override ObjectBuilder.Common.IContainer CreateContainer(ReadOnlySettings settings)
         {
